Extract scores export precedence unit tests into a dedicated module.
Keeps test_scores_exports focused on registry, wire helpers, and integration
queries while precedence matrix tests live beside the pure decision logic.
